// src/enum/privacy.enums.ts
export enum PricayNoticesColumn {
  PrivacyNoticeName = "privacyNoticeName",
  PrivacyNoticeStatusId = "privacyNoticeStatusId",
  PrivacyNoticeStatusName = "privacyNoticeStatusName",
  VersionNumber = "versionNumber",
  PublishedBy = "publishedBy",
  PublishedDate = "publishedDate",
  ModifiedBy = "modifiedBy",
  ModifiedDate = "modifiedDate",
}

export enum Sort {
  Ascending = "asc",
  Descending = "desc",
}